+++ specs/dataBuilder.ts
@@ -34,29 +34,55 @@ import {
     valueColumn,
 } from "../src/columns/columns";
 
-import { getDateRange } from "./helpers";
-
 export class MultiKpiData extends testDataViewBuilder.TestDataViewBuilder {
     public amountOfSeries: number = 5;
 
     public dates: Date[] = [];
     public seriesValues: number[][] = [];
 
-    constructor() {
+    constructor(withMisisngValues?: boolean, brokenMetricIndex: number = 0, howOlderDatesAreInDays: number = 0) {
         super();
 
-        this.dates = getDateRange(
-            new Date(2016, 0, 1),
-            new Date(2016, 0, 10),
-            1000 * 24 * 3600,
-        );
+        const today = new Date();
+        const tommorowOrShifted = new Date(today.getFullYear(), today.getMonth(), today.getDate() + (1 - howOlderDatesAreInDays));
+        const twoWeeksBefore = new Date(tommorowOrShifted.getFullYear(), tommorowOrShifted.getMonth(), tommorowOrShifted.getDate() - 13);
 
-        for (let i: number = 0; i < this.amountOfSeries; i++) {
-            this.seriesValues.push(getRandomNumbers(
-                this.dates.length,
-                -Number.MAX_VALUE,
-                Number.MAX_VALUE,
-            ));
+        // Fill two weeks
+        for (let i = 0; i < 14; i++) {
+            this.dates.push(new Date(twoWeeksBefore.getFullYear(), twoWeeksBefore.getMonth(), twoWeeksBefore.getDate() + i));
+        }
+
+        if (withMisisngValues) {
+            for (let i: number = 0; i < this.amountOfSeries; i++) {
+                if (i === 4) {
+                    const noDataArray: number[] = [];
+                    this.dates.forEach((d) => {
+                        noDataArray.push(undefined);
+                    });
+                    this.seriesValues.push(noDataArray);
+                } else if (i === brokenMetricIndex) {
+                    const valArr: number[] = getRandomNumbers(this.dates.length - 4, -150, 150);
+                    valArr.push(25);
+                    valArr.push(undefined);
+                    valArr.push(undefined);
+                    valArr.push(undefined);
+                    this.seriesValues.push(valArr);
+                } else {
+                    this.seriesValues.push(getRandomNumbers(
+                        this.dates.length,
+                        -150,
+                        150,
+                    ));
+                }
+            }
+        } else {
+            for (let i: number = 0; i < this.amountOfSeries; i++) {
+                this.seriesValues.push(getRandomNumbers(
+                    this.dates.length,
+                    -150,
+                    150,
+                ));
+            }
         }
     }
 
@@ -65,7 +91,7 @@ export class MultiKpiData extends testDataViewBuilder.TestDataViewBuilder {
             source: {
                 displayName: dateColumn.name,
                 format: "%M/%d/yyyy",
-                roles: { Axis: true },
+                roles: { dateColumn: true },
                 type: valueType.ValueType.fromDescriptor({ dateTime: true }),
             },
             values: this.dates,
@@ -76,7 +102,7 @@ export class MultiKpiData extends testDataViewBuilder.TestDataViewBuilder {
                 return {
                     source: {
                         displayName: valueColumn.name,
-                        roles: { Values: true },
+                        roles: { valueColumn: true },
                         type: valueType.ValueType.fromDescriptor({ integer: true }),
                     },
                     values,