+++ src/visualComponent/rootComponent.ts
@@ -209,9 +209,9 @@ export class RootComponent extends BaseContainerComponent<IVisualComponentConstr
             data,
             settings,
         } = options;
-
         this.subtitleComponent.render({
-            dateDifference: data.dateDifference,
+            series: data.series,
+            staleDataDifference: data.staleDateDifference,
             staleDataSettings: settings.staleData,
             subtitleSettings: settings.subtitle,
             warningState: data.warningState,